requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
httpx
debugpy